parser/requirements.txt
lark-parser
pydot